+++ frontend/src/components/GlobalStyles.js
@@ -29,10 +29,3 @@ export const InnerWrapper = styled.section`
     width: 50%;
   }
 `
-
-export const FormSection = styled.section`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-`
